chore(forks): Move bals from Osaka to Amsterdam
