+++ pyproject.toml
@@ -221,7 +221,7 @@ packages = [
     "ethereum.prague.vm.precompiled_contracts.bls12_381",
     "ethereum.osaka",
     "ethereum.osaka.utils",
-    "ethereum.osaka.block_access_lists",
+    "ethereum.amsterdam.block_access_lists",
     "ethereum.osaka.vm",
     "ethereum.osaka.vm.instructions",
     "ethereum.osaka.vm.precompiled_contracts",
