+++ src/ethereum/amsterdam/ssz_types.py
