+++ src/ethereum/amsterdam/vm/instructions/environment.py
@@ -87,6 +87,11 @@ def balance(evm: Evm) -> None:
     # Non-existent accounts default to EMPTY_ACCOUNT, which has balance 0.
     balance = get_account(evm.message.block_env.state, address).balance
 
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_address_access
+
+        track_address_access(evm.message.change_tracker, address)
+
     push(evm.stack, balance)
 
     # PROGRAM COUNTER
@@ -353,6 +358,11 @@ def extcodesize(evm: Evm) -> None:
     # OPERATION
     code = get_account(evm.message.block_env.state, address).code
 
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_address_access
+
+        track_address_access(evm.message.change_tracker, address)
+
     codesize = U256(len(code))
     push(evm.stack, codesize)
 
@@ -394,6 +404,11 @@ def extcodecopy(evm: Evm) -> None:
     # OPERATION
     evm.memory += b"\x00" * extend_memory.expand_by
     code = get_account(evm.message.block_env.state, address).code
+
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_address_access
+
+        track_address_access(evm.message.change_tracker, address)
 
     value = buffer_read(code, code_start_index, size)
     memory_write(evm.memory, memory_start_index, value)
@@ -480,6 +495,11 @@ def extcodehash(evm: Evm) -> None:
 
     # OPERATION
     account = get_account(evm.message.block_env.state, address)
+
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_address_access
+
+        track_address_access(evm.message.change_tracker, address)
 
     if account == EMPTY_ACCOUNT:
         codehash = U256(0)
