+++ src/ethereum/amsterdam/vm/instructions/storage.py
@@ -60,6 +60,16 @@ def sload(evm: Evm) -> None:
         evm.message.block_env.state, evm.message.current_target, key
     )
 
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_storage_read
+
+        track_storage_read(
+            evm.message.change_tracker,
+            evm.message.current_target,
+            key,
+            evm.message.block_env.state,
+        )
+
     push(evm.stack, value)
 
     # PROGRAM COUNTER
@@ -126,6 +136,21 @@ def sstore(evm: Evm) -> None:
     charge_gas(evm, gas_cost)
     if evm.message.is_static:
         raise WriteInStaticContext
+
+    # Track storage write BEFORE modifying state
+    # so we capture the correct pre-value
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_storage_write
+
+        track_storage_write(
+            evm.message.change_tracker,
+            evm.message.current_target,
+            key,
+            new_value,
+            state,
+        )
+
+    # Now modify the storage
     set_storage(state, evm.message.current_target, key, new_value)
 
     # PROGRAM COUNTER
