+++ src/ethereum/amsterdam/vm/instructions/system.py
@@ -108,12 +108,18 @@ def generic_create(
         evm.message.block_env.state, contract_address
     ) or account_has_storage(evm.message.block_env.state, contract_address):
         increment_nonce(
-            evm.message.block_env.state, evm.message.current_target
+            evm.message.block_env.state,
+            evm.message.current_target,
+            evm.message.change_tracker,
         )
         push(evm.stack, U256(0))
         return
 
-    increment_nonce(evm.message.block_env.state, evm.message.current_target)
+    increment_nonce(
+        evm.message.block_env.state,
+        evm.message.current_target,
+        evm.message.change_tracker,
+    )
 
     child_message = Message(
         block_env=evm.message.block_env,
@@ -133,7 +139,14 @@ def generic_create(
         accessed_storage_keys=evm.accessed_storage_keys.copy(),
         disable_precompiles=False,
         parent_evm=evm,
+        change_tracker=evm.message.change_tracker,
     )
+
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_address_access
+
+        track_address_access(evm.message.change_tracker, contract_address)
+
     child_evm = process_create_message(child_message)
 
     if child_evm.error:
@@ -323,7 +336,14 @@ def generic_call(
         accessed_storage_keys=evm.accessed_storage_keys.copy(),
         disable_precompiles=disable_precompiles,
         parent_evm=evm,
+        change_tracker=evm.message.change_tracker,
     )
+
+    if evm.message.change_tracker:
+        from ...block_access_lists.tracker import track_address_access
+
+        track_address_access(evm.message.change_tracker, to)
+
     child_evm = process_message(child_message)
 
     if child_evm.error:
@@ -554,6 +574,7 @@ def selfdestruct(evm: Evm) -> None:
         originator,
         beneficiary,
         originator_balance,
+        evm.message.change_tracker,
     )
 
     # register account for deletion only if it was created
@@ -561,7 +582,12 @@ def selfdestruct(evm: Evm) -> None:
     if originator in evm.message.block_env.state.created_accounts:
         # If beneficiary is the same as originator, then
         # the ether is burnt.
-        set_account_balance(evm.message.block_env.state, originator, U256(0))
+        set_account_balance(
+            evm.message.block_env.state,
+            originator,
+            U256(0),
+            evm.message.change_tracker,
+        )
         evm.accounts_to_delete.add(originator)
 
     # HALT the execution
