+++ src/ethereum/amsterdam/vm/interpreter.py
@@ -190,7 +190,7 @@ def process_create_message(message: Message) -> Evm:
     # added to SELFDESTRUCT by EIP-6780.
     mark_account_created(state, message.current_target)
 
-    increment_nonce(state, message.current_target)
+    increment_nonce(state, message.current_target, message.change_tracker)
     evm = process_message(message)
     if not evm.error:
         contract_code = evm.output
@@ -208,7 +208,12 @@ def process_create_message(message: Message) -> Evm:
             evm.output = b""
             evm.error = error
         else:
-            set_code(state, message.current_target, contract_code)
+            set_code(
+                state,
+                message.current_target,
+                contract_code,
+                message.change_tracker,
+            )
             commit_transaction(state, transient_storage)
     else:
         rollback_transaction(state, transient_storage)
@@ -238,7 +243,13 @@ def process_message(message: Message) -> Evm:
     begin_transaction(state, transient_storage)
 
     if message.should_transfer_value and message.value != 0:
-        move_ether(state, message.caller, message.current_target, message.value)
+        move_ether(
+            state,
+            message.caller,
+            message.current_target,
+            message.value,
+            message.change_tracker,
+        )
 
     evm = execute_code(message)
     if evm.error:
