+++ src/ethereum/osaka/blocks.py
@@ -18,7 +18,6 @@ from ethereum_types.numeric import U64, U256, Uint
 
 from ..crypto.hash import Hash32
 from .fork_types import Address, Bloom, Root
-from .rlp_types import BlockAccessList
 from .transactions import (
     AccessListTransaction,
     BlobTransaction,
@@ -241,14 +240,6 @@ class Header:
     [SHA2-256]: https://en.wikipedia.org/wiki/SHA-2
     """
 
-    bal_hash: Hash32
-    """
-    Hash of the Block Access List containing all accounts and storage
-    locations accessed during block execution. Introduced in [EIP-7928].
-
-    [EIP-7928]: https://eips.ethereum.org/EIPS/eip-7928
-    """
-
 
 @slotted_freezable
 @dataclass
@@ -300,14 +291,6 @@ class Block:
     withdrawals: Tuple[Withdrawal, ...]
     """
     A tuple of withdrawals processed in this block.
-    """
-
-    block_access_list: BlockAccessList
-    """
-    Block Access List containing all accounts and storage locations accessed
-    during block execution. Introduced in [EIP-7928].
-
-    [EIP-7928]: https://eips.ethereum.org/EIPS/eip-7928
     """
