+++ src/ethereum/osaka/fork.py
@@ -30,13 +30,6 @@ from ethereum.exceptions import (
 )
 
 from . import vm
-from .block_access_lists import (
-    StateChangeTracker,
-    build,
-    compute_block_access_list_hash,
-    set_transaction_index,
-    track_balance_change,
-)
 from .blocks import Block, Header, Log, Receipt, Withdrawal, encode_receipt
 from .bloom import logs_bloom
 from .exceptions import (
@@ -251,12 +244,6 @@ def state_transition(chain: BlockChain, block: Block) -> None:
     withdrawals_root = root(block_output.withdrawals_trie)
     requests_hash = compute_requests_hash(block_output.requests)
 
-    # Build and validate Block Access List
-    computed_block_access_list = build(block_output.block_access_list_builder)
-    computed_block_access_list_hash = compute_block_access_list_hash(
-        computed_block_access_list
-    )
-
     if block_output.block_gas_used != block.header.gas_used:
         raise InvalidBlock(
             f"{block_output.block_gas_used} != {block.header.gas_used}"
@@ -274,8 +261,6 @@ def state_transition(chain: BlockChain, block: Block) -> None:
     if block_output.blob_gas_used != block.header.blob_gas_used:
         raise InvalidBlock
     if requests_hash != block.header.requests_hash:
-        raise InvalidBlock
-    if computed_block_access_list_hash != block.header.bal_hash:
         raise InvalidBlock
 
     chain.blocks.append(block)
@@ -596,7 +581,6 @@ def process_system_transaction(
     target_address: Address,
     system_contract_code: Bytes,
     data: Bytes,
-    change_tracker: Optional[StateChangeTracker] = None,
 ) -> MessageCallOutput:
     """
     Process a system transaction with the given code.
@@ -615,8 +599,6 @@ def process_system_transaction(
         Code of the contract to call.
     data :
         Data to pass to the contract.
-    change_tracker :
-        The state change tracker instance.
 
     Returns
     -------
@@ -654,7 +636,6 @@ def process_system_transaction(
         accessed_storage_keys=set(),
         disable_precompiles=False,
         parent_evm=None,
-        change_tracker=change_tracker,
     )
 
     system_tx_output = process_message_call(system_tx_message)
@@ -666,7 +647,6 @@ def process_checked_system_transaction(
     block_env: vm.BlockEnvironment,
     target_address: Address,
     data: Bytes,
-    change_tracker: Optional[StateChangeTracker] = None,
 ) -> MessageCallOutput:
     """
     Process a system transaction and raise an error if the contract does not
@@ -680,8 +660,6 @@ def process_checked_system_transaction(
         Address of the contract to call.
     data :
         Data to pass to the contract.
-    change_tracker :
-        The state change tracker instance.
 
     Returns
     -------
@@ -701,7 +679,6 @@ def process_checked_system_transaction(
         target_address,
         system_contract_code,
         data,
-        change_tracker,
     )
 
     if system_tx_output.error:
@@ -717,7 +694,6 @@ def process_unchecked_system_transaction(
     block_env: vm.BlockEnvironment,
     target_address: Address,
     data: Bytes,
-    change_tracker: Optional[StateChangeTracker] = None,
 ) -> MessageCallOutput:
     """
     Process a system transaction without checking if the contract contains code
@@ -731,8 +707,6 @@ def process_unchecked_system_transaction(
         Address of the contract to call.
     data :
         Data to pass to the contract.
-    change_tracker :
-        The state change tracker instance.
 
     Returns
     -------
@@ -745,7 +719,6 @@ def process_unchecked_system_transaction(
         target_address,
         system_contract_code,
         data,
-        change_tracker,
     )
 
 
@@ -780,44 +753,26 @@ def apply_body(
     """
     block_output = vm.BlockOutput()
 
-    # Initialize Block Access List state change tracker
-    change_tracker = StateChangeTracker(block_output.block_access_list_builder)
-
-    # Set system transaction index for pre-execution system contracts
-    # EIP-7928: System contracts use bal_index 0
-    set_transaction_index(change_tracker, 0)
-
     process_unchecked_system_transaction(
         block_env=block_env,
         target_address=BEACON_ROOTS_ADDRESS,
         data=block_env.parent_beacon_block_root,
-        change_tracker=change_tracker,
     )
 
     process_unchecked_system_transaction(
         block_env=block_env,
         target_address=HISTORY_STORAGE_ADDRESS,
         data=block_env.block_hashes[-1],  # The parent hash
-        change_tracker=change_tracker,
     )
 
-    # EIP-7928: Transactions use bal_index 1 to len(transactions)
     for i, tx in enumerate(map(decode_transaction, transactions)):
-        set_transaction_index(change_tracker, i + 1)
-        process_transaction(
-            block_env, block_output, tx, Uint(i), change_tracker
-        )
+        process_transaction(block_env, block_output, tx, Uint(i))
 
-    # EIP-7928: Post-execution uses bal_index len(transactions) + 1
-    post_execution_index = len(transactions) + 1
-    set_transaction_index(change_tracker, post_execution_index)
-
-    process_withdrawals(block_env, block_output, withdrawals, change_tracker)
+    process_withdrawals(block_env, block_output, withdrawals)
 
     process_general_purpose_requests(
         block_env=block_env,
         block_output=block_output,
-        change_tracker=change_tracker,
     )
 
     return block_output
@@ -826,7 +781,6 @@ def apply_body(
 def process_general_purpose_requests(
     block_env: vm.BlockEnvironment,
     block_output: vm.BlockOutput,
-    change_tracker: StateChangeTracker,
 ) -> None:
     """
     Process all the requests in the block.
@@ -837,8 +791,6 @@ def process_general_purpose_requests(
         The execution environment for the Block.
     block_output :
         The block output for the current block.
-    change_tracker :
-        The state change tracker instance.
     """
     # Requests are to be in ascending order of request type
     deposit_requests = parse_deposit_requests(block_output)
@@ -850,7 +802,6 @@ def process_general_purpose_requests(
         block_env=block_env,
         target_address=WITHDRAWAL_REQUEST_PREDEPLOY_ADDRESS,
         data=b"",
-        change_tracker=change_tracker,
     )
 
     if len(system_withdrawal_tx_output.return_data) > 0:
@@ -862,7 +813,6 @@ def process_general_purpose_requests(
         block_env=block_env,
         target_address=CONSOLIDATION_REQUEST_PREDEPLOY_ADDRESS,
         data=b"",
-        change_tracker=change_tracker,
     )
 
     if len(system_consolidation_tx_output.return_data) > 0:
@@ -877,7 +827,6 @@ def process_transaction(
     block_output: vm.BlockOutput,
     tx: Transaction,
     index: Uint,
-    change_tracker: StateChangeTracker,
 ) -> None:
     """
     Execute a transaction against the provided environment.
@@ -901,8 +850,6 @@ def process_transaction(
         Transaction to execute.
     index:
         Index of the transaction in the block.
-    change_tracker :
-        The state change tracker instance.
     """
     trie_set(
         block_output.transactions_trie,
@@ -933,16 +880,13 @@ def process_transaction(
     effective_gas_fee = tx.gas * effective_gas_price
 
     gas = tx.gas - intrinsic_gas
-    increment_nonce(block_env.state, sender, change_tracker)
+    increment_nonce(block_env.state, sender)
 
     sender_balance_after_gas_fee = (
         Uint(sender_account.balance) - effective_gas_fee - blob_gas_fee
     )
     set_account_balance(
-        block_env.state,
-        sender,
-        U256(sender_balance_after_gas_fee),
-        change_tracker,
+        block_env.state, sender, U256(sender_balance_after_gas_fee)
     )
 
     access_list_addresses = set()
@@ -980,7 +924,6 @@ def process_transaction(
     )
 
     message = prepare_message(block_env, tx_env, tx)
-    message.change_tracker = change_tracker
 
     tx_output = process_message_call(message)
 
@@ -1009,9 +952,7 @@ def process_transaction(
     sender_balance_after_refund = get_account(
         block_env.state, sender
     ).balance + U256(gas_refund_amount)
-    set_account_balance(
-        block_env.state, sender, sender_balance_after_refund, change_tracker
-    )
+    set_account_balance(block_env.state, sender, sender_balance_after_refund)
 
     # transfer miner fees
     coinbase_balance_after_mining_fee = get_account(
@@ -1022,7 +963,6 @@ def process_transaction(
             block_env.state,
             block_env.coinbase,
             coinbase_balance_after_mining_fee,
-            change_tracker,
         )
     elif account_exists_and_is_empty(block_env.state, block_env.coinbase):
         destroy_account(block_env.state, block_env.coinbase)
@@ -1053,7 +993,6 @@ def process_withdrawals(
     block_env: vm.BlockEnvironment,
     block_output: vm.BlockOutput,
     withdrawals: Tuple[Withdrawal, ...],
-    change_tracker: StateChangeTracker,
 ) -> None:
     """
     Increase the balance of the withdrawing account.
@@ -1070,11 +1009,6 @@ def process_withdrawals(
         )
 
         modify_state(block_env.state, wd.address, increase_recipient_balance)
-
-        # Track balance change for BAL
-        # (withdrawals are tracked as system contract changes)
-        new_balance = get_account(block_env.state, wd.address).balance
-        track_balance_change(change_tracker, wd.address, U256(new_balance))
 
         if account_exists_and_is_empty(block_env.state, wd.address):
             destroy_account(block_env.state, wd.address)
