+++ src/ethereum/osaka/vm/__init__.py
@@ -14,7 +14,7 @@ The abstract computer which runs the code stored in an
 """
 
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING, List, Optional, Set, Tuple
+from typing import List, Optional, Set, Tuple
 
 from ethereum_types.bytes import Bytes, Bytes0, Bytes32
 from ethereum_types.numeric import U64, U256, Uint
@@ -22,15 +22,11 @@ from ethereum_types.numeric import U64, U256, Uint
 from ethereum.crypto.hash import Hash32
 from ethereum.exceptions import EthereumException
 
-from ..block_access_lists import BlockAccessListBuilder
 from ..blocks import Log, Receipt, Withdrawal
 from ..fork_types import Address, Authorization, VersionedHash
 from ..state import State, TransientStorage
 from ..transactions import LegacyTransaction
 from ..trie import Trie
-
-if TYPE_CHECKING:
-    from ..block_access_lists import StateChangeTracker
 
 __all__ = ("Environment", "Evm", "Message")
 
@@ -94,9 +90,6 @@ class BlockOutput:
     )
     blob_gas_used: U64 = U64(0)
     requests: List[Bytes] = field(default_factory=list)
-    block_access_list_builder: BlockAccessListBuilder = field(
-        default_factory=BlockAccessListBuilder
-    )
 
 
 @dataclass
@@ -140,7 +133,6 @@ class Message:
     accessed_storage_keys: Set[Tuple[Address, Bytes32]]
     disable_precompiles: bool
     parent_evm: Optional["Evm"]
-    change_tracker: Optional["StateChangeTracker"] = None
 
 
 @dataclass
