+++ src/ethereum/osaka/vm/eoa_delegation.py
@@ -195,8 +195,9 @@ def set_delegation(message: Message) -> U256:
             code_to_set = b""
         else:
             code_to_set = EOA_DELEGATION_MARKER + auth.address
-        set_code(state, authority, code_to_set, message.change_tracker)
-        increment_nonce(state, authority, message.change_tracker)
+        set_code(state, authority, code_to_set)
+
+        increment_nonce(state, authority)
 
     if message.code_address is None:
         raise InvalidBlock("Invalid type 4 transaction: no target")
