+++ src/ethereum/osaka/vm/instructions/system.py
@@ -108,18 +108,12 @@ def generic_create(
         evm.message.block_env.state, contract_address
     ) or account_has_storage(evm.message.block_env.state, contract_address):
         increment_nonce(
-            evm.message.block_env.state,
-            evm.message.current_target,
-            evm.message.change_tracker,
+            evm.message.block_env.state, evm.message.current_target
         )
         push(evm.stack, U256(0))
         return
 
-    increment_nonce(
-        evm.message.block_env.state,
-        evm.message.current_target,
-        evm.message.change_tracker,
-    )
+    increment_nonce(evm.message.block_env.state, evm.message.current_target)
 
     child_message = Message(
         block_env=evm.message.block_env,
@@ -139,14 +133,7 @@ def generic_create(
         accessed_storage_keys=evm.accessed_storage_keys.copy(),
         disable_precompiles=False,
         parent_evm=evm,
-        change_tracker=evm.message.change_tracker,
     )
-
-    if evm.message.change_tracker:
-        from ...block_access_lists.tracker import track_address_access
-
-        track_address_access(evm.message.change_tracker, contract_address)
-
     child_evm = process_create_message(child_message)
 
     if child_evm.error:
@@ -336,14 +323,7 @@ def generic_call(
         accessed_storage_keys=evm.accessed_storage_keys.copy(),
         disable_precompiles=disable_precompiles,
         parent_evm=evm,
-        change_tracker=evm.message.change_tracker,
     )
-
-    if evm.message.change_tracker:
-        from ...block_access_lists.tracker import track_address_access
-
-        track_address_access(evm.message.change_tracker, to)
-
     child_evm = process_message(child_message)
 
     if child_evm.error:
@@ -574,7 +554,6 @@ def selfdestruct(evm: Evm) -> None:
         originator,
         beneficiary,
         originator_balance,
-        evm.message.change_tracker,
     )
 
     # register account for deletion only if it was created
@@ -582,12 +561,7 @@ def selfdestruct(evm: Evm) -> None:
     if originator in evm.message.block_env.state.created_accounts:
         # If beneficiary is the same as originator, then
         # the ether is burnt.
-        set_account_balance(
-            evm.message.block_env.state,
-            originator,
-            U256(0),
-            evm.message.change_tracker,
-        )
+        set_account_balance(evm.message.block_env.state, originator, U256(0))
         evm.accounts_to_delete.add(originator)
 
     # HALT the execution
